Changes aimed at AD's review of the formal appeal (#533)
* WRT the holder to Verifiers text, I would suggest ‘Presents SD-JWT or SD-JWT+KB including selected Disclosures’ to account for the KB-JWT.

* move the "for those who celebrate" part to the end of the sentence and place it in a parenthetical to make it easier to read

* *Lines 253-256: I would reorder the sentences in the definition where the
last sentence is ‘The details of how a JWT for proving Key Binding is
defined in Section 4.3.’  (putting the second sentence first).  Certainly,
details and requirements of how this is done is not suitable for a
terminology section.

* *Lines 248-252: I agree that ‘legitimate possession’ isn’t quite right (nor
provable).  I suggest removing the word ‘legitimate’.  As for implications
that the construction of the system w/ one or more end-users that utilize a
single holder is out of scope.  This draft is merely the format, not the
implementation.

* *Line 215: Still Introduction, the authors could add the word ‘allow’,
since a format itself does not prove possession of the private key.

* *Line 214: I don’t think ‘facility’ here means ‘a physical place’.  I do
think the Authors could pick a better word.  Maybe ‘ability’?

* *Lines 318-320:  This can be easily clarified by changing ‘Holder as part
of the SD-JWT’ to ‘Holder with the SD-JWT’

* Add a sec consideration section to say that key generation, handling, storage, lifecycle needs to be secure.  And that it is out of scope for this draft.

* add doc history for -15 and changes from https://mailarchive.ietf.org/arch/msg/oauth/yJvG93VtP25OcpsoM21ibIy7IrY/

* fix nit per AD

* definition of KB JWT changed per @Sakurann suggestion

Co-authored-by: Kristina <52878547+Sakurann@users.noreply.github.com>

---------

Co-authored-by: Kristina <52878547+Sakurann@users.noreply.github.com>